+++ clippy_lints/src/declared_lints.rs
@@ -193,6 +193,7 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::guidelines::MEM_UNSAFE_FUNCTIONS_INFO,
     crate::guidelines::PASSING_STRING_TO_C_FUNCTIONS_INFO,
     crate::guidelines::UNTRUSTED_LIB_LOADING_INFO,
+    crate::guidelines::UNSAFE_BLOCK_IN_PROC_MACRO_INFO,
     crate::if_let_mutex::IF_LET_MUTEX_INFO,
     crate::if_not_else::IF_NOT_ELSE_INFO,
     crate::if_then_some_else_none::IF_THEN_SOME_ELSE_NONE_INFO,
@@ -614,7 +615,6 @@ pub(crate) static LINTS: &[&crate::LintInfo] = &[
     crate::unnecessary_self_imports::UNNECESSARY_SELF_IMPORTS_INFO,
     crate::unnecessary_wraps::UNNECESSARY_WRAPS_INFO,
     crate::unnested_or_patterns::UNNESTED_OR_PATTERNS_INFO,
-    crate::unsafe_block_in_proc_macro::UNSAFE_BLOCK_IN_PROC_MACRO_INFO,
     crate::unsafe_removed_from_name::UNSAFE_REMOVED_FROM_NAME_INFO,
     crate::unused_async::UNUSED_ASYNC_INFO,
     crate::unused_io_amount::UNUSED_IO_AMOUNT_INFO,
